#466: Proposed changes when ordering based on a dependency graph. (#473)
Fixes #466

+++ src/Orchard.Environment.Extensions.Abstractions/Features/IFeatureInfo.cs
@@ -4,7 +4,7 @@
     {
         string Id { get; }
         string Name { get; }
-        double Priority { get; }
+        int Priority { get; }
         string Category { get; }
         string Description { get; }
         IExtensionInfo Extension { get; }
